feat: 将课程相关ID统一改为UUID类型
- 删除了旧的Generator.kt文件
- 将所有课程相关的ID类型从Int或Long改为UUID，确保ID的唯一性和一致性
- 修改了CourseDao接口，插入和批量插入方法返回UUID
- 更新CourseRepositoryImpl中课程及课程节点的ID处理，使用UUID
- 修改了相关领域模型（User、Table、Course、OrdinarySchedule、GlobalTableSetting、TimeSlot等）中ID的默认值和类型
- 调整了多个UseCase、ViewModel、UI层和导航参数，支持UUID类型的ID传递和处理
- 修正了同步消息日志打印，使用UUID标识
- 更新了课程解析器Parser，使用UUID生成和查找课程ID
- 统一了代码中UUID的导入和使用，替换了原有的Int/Long ID逻辑
- 修复了部分逻辑判断中ID的默认值判断，使用EMPTY_UUID常量替代-1或0
- 优化了部分UI组件和状态管理，支持UUID类型的ID
- 相关常量和扩展函数支持UUID转换和处理

此次提交完成了课程ID体系的UUID化改造，提升了系统的扩展性和数据一致性。

+++ app/src/main/java/com/example/todoschedule/ui/auth/LoginViewModel.kt
@@ -51,7 +51,7 @@ class LoginViewModel @Inject constructor(
 
             result.onSuccess { user ->
                 // 登录成功，保存用户会话
-                saveLoginSessionUseCase(user.id.toLong(), user.token) // 保存用户ID和token
+                saveLoginSessionUseCase(user.id, user.token) // 保存用户ID和token
 
                 // 触发立即同步，使用应用级别的协程作用域
                 applicationScope.launch {
